bert_vits2/requirements.txt
style-bert-vits2
PySoundFile
fastapi
uvicorn[standard]
